Carving out common code out of HTTPMessage.swift
